+++ src/components/Sidebar/index.ts
@@ -1,0 +1,6 @@
+import RvSidebar from "./RvSidebar";
+import RvMenu from "./RvMenu";
+import RvMenuItem from "./RvMenuItem";
+import RvSubMenu from "./RvSubMenu";
+
+export { RvSidebar, RvMenu, RvMenuItem, RvSubMenu };