+++ src/components/Span/index.ts
@@ -1,3 +1,4 @@
 import RvSpan from "./RvSpan";
 
 export { RvSpan };
+export * from "./style";