+++ src/components/index.ts
@@ -1,5 +1,6 @@
 export * from "./Box";
 export * from "./Button";
 export * from "./Card";
+export * from "./Sidebar";
 export * from "./Span";
 export * from "./config";