+++ src/types/components/index.ts
@@ -1,4 +1,5 @@
 export * from "./Div";
 export * from "./Button";
 export * from "./Card";
+export * from "./Sidebar";
 export * from "./Span";